swiftvalidator/src/net/uglobal/swiftvalidator/swift/field/Field97A.java

package net.uglobal.swiftvalidator.swift.field;



/**
 * Field format : (qualifier)(accountNumber)
 * 
 * AUTO GENERATED - DO NOT HAND MODIFY
 * Generated on Sun Dec 09 17:22:28 EST 2012
 * 
 */
public class Field97A
    extends net.uglobal.swiftvalidator.swift.field.Field
{


    public Field97A(String validateData, net.uglobal.swiftvalidator.xml.beans.Field xmlField) {
        super(validateData, xmlField);
    }

    @Override
    protected void parse() {
        addComponents(net.uglobal.swiftvalidator.swift.field.FieldComponentParseUtil.parse(validateData, getContentFormat()));
    }

    public String getQualifier() {
        return getComponent1();
    }

    public String getAccountNumber() {
        return getComponent2();
    }

    public String toString() {
        return (new StringBuilder()).append("[" ).append( "qualifier=" + getComponent1()).append( ", accountNumber=" + getComponent2() ).append( "]").toString();
    }

}
